// src/interfaces/db.interfaz.ts
import * as sql from 'mssql';

export interface IParametro{
    nombre: string,
    tipo:
        | sql.ISqlTypeFactoryWithLength
        | sql.ISqlTypeFactoryWithNoParams
        | sql.ISqlTypeFactoryWithPrecisionScale
        | sql.ISqlTypeFactoryWithScale
        | sql.ISqlTypeFactoryWithTvpType
    valor: any
}